ENH: Make the remaining integer conversions explicit
The -Wsign-conversion findings that are not allocation, copy or I/O
lengths: file offsets and positions, extension bookkeeping, and a few
conversions that were wrong rather than merely implicit.

  File offsets.  znztell() returns a signed znz_off_t stored in size_t
  locals; the znzseek() arguments are cast to znz_off_t so the arithmetic
  is done at the width the function takes.

  fslio's nvox.  The product of seven int dimensions assigned to a size_t
  field, widened per operand so it is computed at 64 bits.

  The ANALYZE orientation byte.  nifti_convert_nhdr2nim() read it with
  `unsigned char c = *((char *)(&nhdr.qform_code));`, taking a signed
  char and widening it into an unsigned one, so any value above 127
  arrived sign-extended.  Reading through unsigned char * gives the byte
  that is actually in the header.

  expat's XML_Parse takes an int length while blen is unsigned; the
  conversion is now at the call rather than hidden in the variable.

Lines that upstream PRs also change are left alone, including the
d3matrix and d4matrix allocations in fslio.c, which PR #22 rewrites.

+++ znzlib/znzlib.c
@@ -147,7 +147,7 @@ size_t znzread(void* buf, size_t size, size_t nmemb, znzFile file)
        nread = gzread(file->zfptr, (void *)cbuf, n2read);
        if( nread < 0 ) return nread; /* returns -1 on error */
 
-       remain -= nread;
+       remain -= (size_t)nread;
        cbuf += nread;
 
        /* require reading n2read bytes, so we don't get stuck */
@@ -181,7 +181,7 @@ size_t znzwrite(const void* buf, size_t size, size_t nmemb, znzFile file)
        /* gzread returns 0 on error, but in case that ever changes... */
        if( nwritten < 0 ) return nwritten;
 
-       remain -= nwritten;
+       remain -= (size_t)nwritten;
        cbuf += nwritten;
 
        /* require writing n2write bytes, so we don't get stuck */
